automl/requirements.txt
matplotlib==3.7.5
numpy==1.26.4
pycaret==3.3.2
scikit_learn==1.4.2
scikit_plot==0.3.7
scipy==1.13.1
streamlit==1.30.0
streamlit_option_menu==0.3.2
streamlit_pandas_profiling==0.1.3
ydata_profiling==4.8.3
